add web-component.js & bento-date-display.js
