+++ extensions/amp-date-display/1.0/bento-date-display.js
@@ -1,0 +1,3 @@
+import {defineElement} from './web-component';
+
+defineElement();
